+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Update vercel.
 
+### Changed
+- Suggestions.
+
 
 ## [0.13.0] - 1.12.2025